requirements.txt
torch>=1.9.0
numpy>=1.21.0
requests>=2.26.0
python-dotenv>=0.19.0
